assign new service functions + missing some dao methods + some needs

+++ app/controllers/project/views.py
@@ -411,6 +411,42 @@ def sampleusers(project_name, sample_name):
 	return resp
 
 
+@project.route('/sample/<role>/add', methods=['POST'])
+def addRole2Sample(role):
+	if not request.json: abort(400)
+	req = request.json
+	samples = {"samples":project_service.get_samples(req['projectname'])}
+	res = {}
+	if 'samplename' in req:
+		if not req['samplename'] in samples["samples"]: abort(404)
+		possible_roles = [x[0] for x in project_service.get_possible_roles()]
+		roleInt = [r[0] for r in project_service.get_possible_roles() if r[1] == role][0]
+		user = user_service.get_by_username(req['username'])
+		if not user: abort(400)
+		project_service.add_or_delete_sample_role(user, req['samplename'], req['projectname'], roleInt, False)
+		sample = project_service.get_sample(req['samplename'], req['projectname'], current_user)
+		res = sample
+	js = json.dumps(res)
+	resp = Response(js, status=200,  mimetype='application/json')
+	return resp
+
+@project.route('/sample/<role>/remove', methods=['POST'])
+def removeRole2Sample(role):
+	if not request.json: abort(400)
+	req = request.json
+	print('req', req)
+	samples = {"samples":project_service.get_samples(req['projectname'])}
+	if 'samplename' in req:
+		print('hello')
+		if not req['samplename'] in samples["samples"]: abort(404)
+		possible_roles = [x[0] for x in project_service.get_possible_roles()]
+		roleInt = [r[0] for r in project_service.get_possible_roles() if r[1] == role][0]
+		user = user_service.get_by_username(req['username'])
+		if not user: abort(400)
+		project_service.add_or_delete_sample_role(user, req['samplename'], req['projectname'], roleInt, True)
+	js = json.dumps({})
+	resp = Response(js, status=200,  mimetype='application/json')
+	return resp
 
 
 @project.route('/<project_name>/sample/<sample_name>/users', methods=['POST'])
@@ -439,13 +475,11 @@ def userrole(project_name, sample_name):
 		user = user_service.get_by_id(u)
 		if not user: abort(400)
 		if r:
-			if r not in possible_roles:
-				abort(400)
+			if r not in possible_roles: abort(400)
 			project_service.create_add_sample_role(u, sample_name, project.id, r)
 		else:
 			sr = SampleRole.query.filter_by(projectid=project.id, samplename=sample_name, userid=u).first()
-			if sr:
-				project_service.delete_sample_role(sr)
+			if sr: project_service.delete_sample_role(sr)
 	return sampleusers(project_name, sample_name)
